Replaced switch case with tag dispatching
