+++ commands/List.js
@@ -1,0 +1,13 @@
+function handle({ recommendations }) {
+
+  const listEmbed = {
+    description: [...recommendations.tags].join(', ')
+  }
+
+  return {
+    reply: (msg) => msg.channel.send('**All tags:**', { embed : listEmbed })
+  }
+
+}
+
+module.exports = handle
